agent/requirements.txt
langchain>=1.0,<2.0
langchain-openai>=1.0,<2.0
langchain-community>=0.4,<0.5
langchain-ollama>=1.0,<2.0
langgraph>=1.0,<2.0
asyncpg>=0.31,<1.0
psycopg2-binary>=2.9,<3.0
pypdf>=6.0,<7.0
python-docx>=1.1,<2.0
python-dotenv>=1.0,<2.0
openai>=2.0,<3.0
httpx>=0.27,<1.0
duckduckgo-search>=6.0,<7.0
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
uvicorn[standard]
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
ruff>=0.6,<1.0
hanlp>=1.0,<2.0

# 可选：敏感信息过滤（需要 PyTorch）
# opf>=0.1.0
